Completed all 3 assignments. Code and outputs added.

+++ utils/srt_to_csv.py
@@ -15,22 +15,26 @@ def write_csv(row,path):
         csvFile.close()
 
 def srt_to_csv(file='./videos/DJI_0301.SRT'):
-    path = './files/'+((file.split('/')[-1]).split('.')[0])+'.csv'
-    if(os.path.exists(path)):
-        os.remove(path)
+    if(os.path.exists(file) == False):
+        print("File not found!")
 
-    f = open(file,'r')
-    l = f.read().split('\n\n')
-    l.pop(len(l)-1)
-    dic = {}
+    else:
+        path = './files/'+((file.split('/')[-1]).split('.')[0])+'.csv'
+        if(os.path.exists(path)):
+            os.remove(path)
 
-    for i in l:
-        subl = i.split('\n')
-        time = subl[1]
-        coord = subl[2]
-        start,end = time.split('-->')
-        lon,lat,_ = coord.rstrip().split(',')
-        write_csv([start.strip(' '),end.strip(' '),lat,lon],path)
+        f = open(file,'r')
+        l = f.read().split('\n\n')
+        l.pop(len(l)-1)
+        dic = {}
+
+        for i in l:
+            subl = i.split('\n')
+            time = subl[1]
+            coord = subl[2]
+            start,end = time.split('-->')
+            lon,lat,_ = coord.rstrip().split(',')
+            write_csv([start.strip(' '),end.strip(' '),lat,lon],path)
 
 if __name__ == '__main__':
     srt_to_csv()
